@@ -2,29 +2,31 @@
   "name": "webizen-app-docs",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "dev:turbopack": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "lint:fix": "next lint --fix"
+    "lint:fix": "next lint --fix",
+    "deploy": "gh-pages -d out"
   },
   "dependencies": {
-    "eslint-config-next": "^14.2.30",
-    "next": "^14.2.30",
-    "nextra": "^3.3.1",
-    "nextra-theme-docs": "^1.2.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "^15.0.1",
+    "nextra": "^4.2.17",
+    "nextra-theme-docs": "^4.2.17",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "autoprefixer": "^10.4.20",
-    "eslint": "^8.57.1",
+    "eslint": "^9.12.0",
+    "eslint-config-next": "^15.0.1",
     "eslint-plugin-mdx": "^3.1.4",
     "eslint-plugin-react": "^7.37.1",
+    "gh-pages": "^6.2.0",
     "postcss": "^8.4.47",
-    "tailwindcss": "^3.4.13",
-    "typescript": "^5.8.3"
+    "tailwindcss": "^3.4.13"
   }
 }
