@@ -1,14 +1,9 @@
-/** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
-    './pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './components/**/*.{js,ts,jsx,tsx,mdx}',
-    './app/**/*.{js,ts,jsx,tsx,mdx}',
-    './theme.config.jsx',
-    './node_modules/nextra-theme-docs/dist/**/*.{js,ts,jsx,tsx}',
+    './pages/**/*.{js,jsx,md,mdx}',
+    './components/**/*.{js,jsx}',
+    './theme.config.mjs'
   ],
-  theme: {
-    extend: {},
-  },
-  plugins: [],
+  theme: { extend: {} },
+  plugins: []
 };
