@@ -1,12 +1,11 @@
-import nextra from 'nextra';
-export default nextra({
-  theme: 'nextra-theme-docs',
-  themeConfig: './theme.config.mjs'
-})({
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   output: 'export',
   basePath: '/webizen-app-docs',
   trailingSlash: true,
   images: {
     unoptimized: true
   }
-});
+};
+
+export default nextConfig;
