@@ -3,11 +3,19 @@ export default {
   project: {
     link: 'https://github.com/WebizenAI/webizen-app-docs'
   },
+  docsRepositoryBase: 'https://github.com/WebizenAI/webizen-app-docs/tree/main',
   search: {
     provider: 'pagefind'
   },
   footer: {
-    text: 'WebizenAI © 2025'
+    text: `WebizenAI © ${new Date().getFullYear()}`
+  },
+  sidebar: {
+    defaultMenuCollapseLevel: 1,
+    toggleButton: true
+  },
+  toc: {
+    float: true
   },
   basePath: '/webizen-app-docs'
 };
